# perf: prototype preventing redundant data refetch every swap step