<!-- < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < ☺
v                               ✰  Thanks for creating a PR! ✰    
v    Before smashing the submit button please review the checkboxes.
v    If a checkbox is n/a - please still include it but + a little note why
v    If your PR doesn't close an issue, that's OK!  Just remove the Closes: #XXX line!
☺ > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > >  -->

Closes: #XXX

## What is the purpose of the change

According to latest swap profiles off of `main` at [this commit](https://github.com/osmosis-labs/osmosis/commit/7f0441ab1da0251a312606c50a851eb5ffebfa38), we have bottlenecks in refetching ticks, fee and incentive accumulators redundantly from state.

![image](https://github.com/osmosis-labs/osmosis/assets/34196718/193f5dec-395b-42c3-a957-ef1f95588f0a)

This PR prototypes removing all redundant data refetch. Actual changes will be submitted via separate PRs to isolate each change.

The core of the change is three-fold:
- avoid refetching ticks. They can be retrieved from the store iterator so there is no need to access the store with `GetTickInfo` again when crossing a tick
- fee accumulator is refetched every iteration
- incentives uptime accumulators are refetched every iteration

However, there is no need to refetch these accumulators as `accum` package guarantees that the variable is to be mutated together with state.

The resulting flafe graph from all 3 changes is attached below:
![image](https://github.com/osmosis-labs/osmosis/assets/34196718/03b7ae1d-fbbc-4092-88ae-192d0fbc9330)

Notice 3 results visible on the flame graph:

1. `GetTickInfo` overhead is reduced and not visible
2. `GetUptimeAccumulators` overhead is reduced and not visible
3. `GetFeeAccumulators` overhead is reduced and not visible 


Benchmark with leveldb and 10k positions shows a 40% improvement:
```bash
benchstat current.bench latest.bench
current.bench:5: missing iteration count
latest.bench:5: missing iteration count
goos: linux
goarch: amd64
pkg: github.com/osmosis-labs/osmosis/v15/x/concentrated-liquidity
cpu: Intel(R) Xeon(R) Gold 6248 CPU @ 2.50GHz
                    │ current.bench │            latest.bench             │
                    │    sec/op     │    sec/op     vs base               │
SwapExactAmountIn-4    341.1m ± 14%   204.7m ± 11%  -39.97% (p=0.002 n=6)
```